Revert "Contract parameters moved to SaveEntity."
