+++ src/Api/appsettings.Uat.json
@@ -31,8 +31,7 @@
     "ClientId": "CLIENT_ID",
     "SSoIntegrationService": "SSoIntegrationServiceURL",
     "TsizlUrl": "TSIZL_URL",
-    "ContractUrl": "CONTRACT_URL",
-    "AmorphieWorkflowUrl": "AMORPHIE_WORKFLOW_URL"
+    "ContractUrl": "CONTRACT_URL"
   },
   "ServiceEndpoint": {
     "DMSService": "DMSSERVICE"
