+++ src/Api/appsettings.json
@@ -30,8 +30,7 @@
     "ApiSecret": "API_SECRET",
     "ClientId": "CLIENT_ID",
     "TsizlUrl": "TSIZL_URL",
-    "ContractUrl": "CONTRACT_URL",
-    "AmorphieWorkflowUrl": "AMORPHIE_WORKFLOW_URL"
+    "ContractUrl": "CONTRACT_URL"
   },
   "ServiceEndpoint": {
     "DMSService": "DMSSERVICE"
