+++ workers/Worker.App/Application/Common/Models/StaticValues.cs
@@ -9,6 +9,5 @@ public static class StaticValues
     public static string MessagingGateway { get; set; }
     public static string TsizlUrl { get; set; }
     public static string ContractUrl { get; set; }
-    public static string AmorphieWorkflowUrl { get; set; }
 }
