+++ workers/Worker.App/Application/Workers/Commands/StartFreeContractApprovals/StartFreeContractApprovalCommand.cs
@@ -34,83 +34,86 @@ namespace Worker.App.Application.Workers.Commands.StartFreeContractApprovals
 
         public async Task<Response<StartFreeContractApprovalResponse>> Handle(StartFreeContractApprovalCommand request, CancellationToken cancellationToken)
         {
-            if (String.IsNullOrEmpty(request.ContractCode))
+            using (var client = new HttpClient())
             {
-                var map = _context.ContractMaps.Where(x => x.EndorsementCode == request.ContractTitle).FirstOrDefault();
-                request.ContractCode = map != null ? map.ContractCode : "";
-                request.ToLangCode = map != null ? map.Language : "";
-            }
-
-            if (String.IsNullOrEmpty(request.ContractCode))
-                throw new Exception("ContractCode can not be empty!");
-
-            var orderFreeDocuments = _context.Documents.Where(x => x.OrderId == request.OrderId && x.Type == "PlainText").ToList();
-            List<object> freeDocuments = new List<object>();
-
-            foreach (var orderDocument in orderFreeDocuments)
-            {
-                byte[] decodedBytes = Convert.FromBase64String(orderDocument.Content);
-                string decodedString = Encoding.UTF8.GetString(decodedBytes);
-
-                var documentCode = _context.ContractMaps.Where(x => x.EndorsementCode == orderDocument.Name).Select(x => x.DocumentCode).FirstOrDefault();
-                freeDocuments.Add(new
+                if (String.IsNullOrEmpty(request.ContractCode))
                 {
-                    DocumentCode = documentCode, //ASK: Her bir serbset için tanım yapılacak contract'a template free-document seçilecek.
-                    DocumentContent = decodedString
+                    var map = _context.ContractMaps.Where(x => x.EndorsementCode == request.ContractTitle).FirstOrDefault();
+                    request.ContractCode = map != null ? map.ContractCode : "";
+                    request.ToLangCode = map != null ? map.Language : "";
+                }
+
+                if (String.IsNullOrEmpty(request.ContractCode))
+                    throw new Exception("ContractCode can not be empty!");
+
+                var orderFreeDocuments = _context.Documents.Where(x => x.OrderId == request.OrderId && x.Type == "PlainText").ToList();
+
+                List<object> freeDocuments = new List<object>();
+
+                foreach (var orderDocument in orderFreeDocuments)
+                {
+                    byte[] decodedBytes = Convert.FromBase64String(orderDocument.Content);
+                    string decodedString = Encoding.UTF8.GetString(decodedBytes);
+
+                    var documentCode = _context.ContractMaps.Where(x => x.EndorsementCode == orderDocument.Name).Select(x => x.DocumentCode).FirstOrDefault();
+                    freeDocuments.Add(new {
+                        DocumentCode = documentCode, //ASK: Her bir serbset için tanım yapılacak contract'a template free-document seçilecek.
+                        DocumentContent = decodedString
+                    });
+                }
+
+                client.BaseAddress = new Uri(StaticValues.ContractUrl);
+                // client.BaseAddress = new Uri("https://test-amorphie-workflow.burgan.com.tr");
+
+                object reqObj = new
+                {
+                    ContractCode = request.ContractCode,
+                    ToUserReference = request.ToUserReference,
+                    ToCustomerNo = request.ToCustomerNo,
+                    ToLangCode = request.ToLangCode,
+                    ContractTitle = request.ContractTitle,
+                    SetTimeout = request.SetTimeout,
+                    ToBusinessLine = request.ToBusinessLine,
+                    FreeDocuments = freeDocuments
+                };
+
+                var response = new StartFreeContractApprovalResponse
+                {
+                    WorkflowId = request.ContractInstanceId
+                };
+
+                var json = JsonSerializer.Serialize(reqObj);
+                var content = new StringContent(json, Encoding.UTF8, "application/json");
+
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", request.AuthToken.Replace("Bearer ", ""));
+                var result = await client.PostAsync("workflow/instance/" + response.WorkflowId.ToString() + "/transition/free-contract-approval-start", content);
+                var responseContent = await result.Content.ReadAsStringAsync();
+                
+                if (result.IsSuccessStatusCode)
+                    {
+                        Log.ForContext("ContractInstanceId", request.ContractInstanceId)
+                        .ForContext("FreeContractBody", json)
+                        .ForContext("HttpResponseStatus", result.StatusCode)
+                        .Information($"Free contract started.");
+                    }
+                    else
+                    {
+                        Log.ForContext("ContractInstanceId", request.ContractInstanceId)
+                        .ForContext("UploadedDocument", json)
+                        .ForContext("HttpResponseStatus", result.StatusCode)
+                        .Error($"StartFreeContract Error. Content: " + responseContent);
+                    }
+                
+
+                _context.ContractStarts.Add(new ContractStart
+                {
+                    ContractStartId = Guid.NewGuid(),
+                    ContractInstanceId = response.WorkflowId,
+                    OrderId = Guid.Parse(request.OrderId)
                 });
+
+                return Response<StartFreeContractApprovalResponse>.Success(response, 200);
             }
-            var client = new HttpClient();
-            // client.BaseAddress = new Uri(StaticValues.AmorphieWorkflowUrl);
-            client.BaseAddress = new Uri("https://preprod-amorphie-workflow.burgan.com.tr/");
-            object reqObj = new
-            {
-                ContractCode = request.ContractCode,
-                ToUserReference = request.ToUserReference,
-                ToCustomerNo = request.ToCustomerNo,
-                ToLangCode = request.ToLangCode,
-                ContractTitle = request.ContractTitle,
-                SetTimeout = request.SetTimeout,
-                ToBusinessLine = request.ToBusinessLine,
-                FreeDocuments = freeDocuments
-            };
-
-            var response = new StartFreeContractApprovalResponse
-            {
-                WorkflowId = request.ContractInstanceId
-            };
-
-            var json = JsonSerializer.Serialize(reqObj);
-            var content = new StringContent(json, Encoding.UTF8, "application/json");
-
-            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", request.AuthToken.Replace("Bearer ", ""));
-            client.DefaultRequestHeaders.Add("User", "");
-            client.DefaultRequestHeaders.Add("Behalf-Of-User", "");
-            var result = await client.PostAsync("workflow/instance/" + response.WorkflowId.ToString() + "/transition/free-contract-approval-start", content);
-            var responseContent = await result.Content.ReadAsStringAsync();
-
-            if (result.IsSuccessStatusCode)
-            {
-                Log.ForContext("ContractInstanceId", request.ContractInstanceId)
-                .ForContext("FreeContractBody", json)
-                .ForContext("HttpResponseStatus", result.StatusCode)
-                .Information($"Free contract started.");
-            }
-            else
-            {
-                Log.ForContext("ContractInstanceId", request.ContractInstanceId)
-                .ForContext("UploadedDocument", json)
-                .ForContext("HttpResponseStatus", result.StatusCode)
-                .Error($"StartFreeContract Error. Content: " + responseContent);
-            }
-
-            _context.ContractStarts.Add(new ContractStart
-            {
-                ContractStartId = Guid.NewGuid(),
-                ContractInstanceId = response.WorkflowId,
-                OrderId = Guid.Parse(request.OrderId)
-            });
-
-            return Response<StartFreeContractApprovalResponse>.Success(response, 200);
         }
     }
 }
