+++ workers/Worker.App/Application/Workers/Commands/UploadContractDocumentInstances/UploadContractDocumentInstanceCommand.cs
@@ -18,9 +18,6 @@ namespace Worker.App.Application.Workers.Commands.UploadContractDocumentInstance
     {
         public string OrderId { get; set; }
         public string AuthToken { get; set; }
-        public string ContractInstanceId { get; set; }
-        public string ContractCode { get; set; }
-        public string Language { get; set; }
         public string ToBusinessLine { get; set; }
         public string ToUserReference { get; set; }
     }
@@ -36,16 +33,53 @@ namespace Worker.App.Application.Workers.Commands.UploadContractDocumentInstance
 
         public async Task<Response<UploadContractDocumentInstanceResponse>> Handle(UploadContractDocumentInstanceCommand request, CancellationToken cancellationToken)
         {
-            Log.Information($"UploadContractDocumentInstanceCommand Started.");
-            
+            Guid contractInstanceId = Guid.NewGuid();
+            Log.ForContext("ContractInstanceId", contractInstanceId).Information($"UploadContractDocumentInstanceCommand Started.");
             var orderDocuments = _context.Documents.Where(x => x.OrderId == request.OrderId).ToList();
-            var currentContract = _context.ContractMaps.Where(x => x.ContractCode == request.ContractCode)
-                .GroupBy(x => x.ContractCode)
-                .Select(g => new
+            var documentNames = orderDocuments.Select(x => x.Name.Replace(".pdf", "")).ToList();
+
+            var maps = _context.ContractMaps.Where(x => documentNames.Contains(x.EndorsementCode) && x.RequiresFullMatch)
+            .GroupBy(x => x.ContractCode)
+            .Select(g => new
+            {
+                ContractCode = g.Key,
+                Items = g.ToList()
+            }).ToDictionary(x => x.ContractCode, x => x.Items); //ASK: serbest döküman için ContractCode boş bırakma?
+
+            var currentContract = new KeyValuePair<string, List<ContractMap>>();
+
+            foreach (var contract in maps)
+            {
+                var fullMatchList = contract.Value.Select(x => x.EndorsementCode).ToList();
+                bool areEqual = new HashSet<string>(fullMatchList).SetEquals(documentNames);
+                if (areEqual)
+                {
+                    currentContract = contract;
+                    Log.ForContext("ContractInstanceId", contractInstanceId).ForContext("ContractCode", currentContract.Key).Information($"UploadContractDocumentInstanceCommand Contract Found From RequiresFullMatch.");
+                    break;
+                }
+            }
+
+            if (String.IsNullOrEmpty(currentContract.Key))
+            {
+                currentContract = _context.ContractMaps.Where(x => documentNames.Contains(x.EndorsementCode) && !x.RequiresFullMatch)
+                .GroupBy(x => x.ContractCode).Select(g => new
                 {
                     ContractCode = g.Key,
                     Items = g.ToList()
                 }).ToDictionary(x => x.ContractCode, x => x.Items).FirstOrDefault();
+                Log.ForContext("ContractInstanceId", contractInstanceId).ForContext("ContractCode", currentContract.Key).Information($"UploadContractDocumentInstanceCommand Contract Found From Not RequiresFullMatch.");
+            }
+
+            if (String.IsNullOrEmpty(currentContract.Key))
+            {
+                throw new Exception("ContractCode not found!");
+            }
+
+            var config = _context.Configs.FirstOrDefault(x => x.OrderId == request.OrderId);
+            config.ContractParameters = contractInstanceId + ";" + currentContract.Key + ";tr-TR";
+            _context.Configs.Update(config);
+            _context.SaveChanges();
 
             var client = new HttpClient();
             client.BaseAddress = new Uri(StaticValues.ContractUrl);
@@ -57,7 +91,7 @@ namespace Worker.App.Application.Workers.Commands.UploadContractDocumentInstance
                     UploadContractDocumentInstanceModel uploadDoc = new UploadContractDocumentInstanceModel
                     {
                         ContractCode = currentContract.Key,
-                        ContractInstanceId = Guid.Parse(request.ContractInstanceId),
+                        ContractInstanceId = contractInstanceId,
                     };
 
                     uploadDoc.DocumentInstanceId = Guid.NewGuid();
@@ -81,14 +115,14 @@ namespace Worker.App.Application.Workers.Commands.UploadContractDocumentInstance
                     var result = await client.PostAsync("document/uploadInstance", content);
                     if (result.IsSuccessStatusCode)
                     {
-                        Log.ForContext("ContractInstanceId", request.ContractInstanceId)
+                        Log.ForContext("ContractInstanceId", contractInstanceId)
                         .ForContext("UploadedDocument", json)
                         .ForContext("HttpResponseStatus", result.StatusCode)
                         .Information($"UploadContractDocumentInstanceCommand Document Uploaded.");
                     }
                     else
                     {
-                        Log.ForContext("ContractInstanceId", request.ContractInstanceId)
+                        Log.ForContext("ContractInstanceId", contractInstanceId)
                         .ForContext("UploadedDocument", json)
                         .ForContext("HttpResponseStatus", result.StatusCode)
                         .Error($"UploadContractDocumentInstanceCommand Document Upload Error.");
@@ -99,7 +133,7 @@ namespace Worker.App.Application.Workers.Commands.UploadContractDocumentInstance
 
             var response = new UploadContractDocumentInstanceResponse
             {
-                ContractInstanceId = request.ContractInstanceId,
+                ContractInstanceId = contractInstanceId,
                 ContractCode = currentContract.Key,
                 Language = "tr-TR"
             };
