+++ workers/Worker.App/appsettings.Preprod.json
@@ -27,8 +27,7 @@
     "Sso": "SSO",
     "MessagingGateway": "MESSAGING_GATEWAY",
     "TsizlUrl": "TSIZL_URL",
-    "ContractUrl": "CONTRACT_URL",
-    "AmorphieWorkflowUrl": "AMORPHIE_WORKFLOW_URL"
+    "ContractUrl": "CONTRACT_URL"
   },
   "ServiceEndpoint": {
     "DMSService": "DMSSERVICE"
